add email login edge cases and separate files for login
